Compound v2.2 Fixes Post-Audit
This patch addresses the core issues from the OpenZeppelin preliminary audit. We change the maximim borrow rate mantissa, we change how ownership of pause guardian works, we require the admin to unpause the protocol, we decide to check and return error codes from enter-market-on-borrow among a few other minor changes and cleanups.