Modify DAIInterestRateModel.sol => DAIInterestRateModelV2.sol
The primary intention behind this change is to "smooth out" the cDai interest rate model to promote greater rate stability.

It does so by increasing the `gapPerBlock`, or additional "slope" of the interest rate curve added to the Dai stability fee, from 0.05% to 2%.

There are also a few minor changes, designed to clean up the interest rate model contract.

Summary of changes:
- `gapPerBlock` constant changed from `0.05e16 / blocksPerYear` to `2e16 / blocksPerYear`
- `uint` => `uint256` (explicit > implicit)
- added parenthesis to SafeMath expressions to make ordering explicit
- used an interface (rather than an abstract contract) for the maker interfaces
- modified a tiny bit of `poke()` to add `gapPerBlock` to `stabilityFeePerBlock` _before_ subtracting `baseRatePerBlock` for resiliency against a very unlikely underflow
- named some unnamed return values