All markets bugfix; add scens (#60)
This patch tracks all markets instead of just COMP markets, since we need to be able to iterate over the superset of markets in most cases, to ensure that toggling COMP-enabled on markets always clears out speeds and re-initializes state correctly. The initial upgrade should take all the existing markets as a parameter, and new markets are initialized via the call to `_supportMarket`.