Add scenario changes and a forking simulation; remove _becomeG3
It will be easier for someone to make a proposal if they don't need to worry about the _becomeG3 arguments.
The proposal may now be made, as in the forking simulation.

We also fix up the test harnesses for the new Comptroller impl generation, and add a storage check for it.